+++ converter.html
@@ -2,7 +2,7 @@
 <html>
   <head>
     <title>Unit Converter</title>
-    <script src="./converter.js"></script>
+    <script type="module" src="./converter.js"></script>
   </head>
   <body>
     <section>
